Moved/renamed level parser file
